+++ controllers/DatasetController.js
@@ -59,7 +59,6 @@ class DatasetController {
         }
 
         catch (error) {
-            console.log(error)
             return res.status(500).json({ msg: statusMessages.connectionError })
         }
     }